+++ lib/widgets/home/stats/your_stats.dart
@@ -6,7 +6,7 @@ import 'package:provider/provider.dart';
 import 'package:routinger/constants/layout.dart';
 
 //Provider Imports
-import '../../provider/experience_points.dart';
+import '../../../provider/experience_points.dart';
 
 class YourStats extends StatelessWidget {
   @override
