+++ lib/widgets/home/taskslist/scheduled_card.dart
@@ -7,19 +7,19 @@ import 'package:intl/intl.dart';
 import 'package:mdi/mdi.dart';
 
 // Constant Imports
-import '../../constants/enums.dart';
+import '../../../constants/enums.dart';
 
 // Helper Imports
-import '../../helper/db_helper.dart';
+import '../../../helper/db_helper.dart';
 
 // Provider Imports
-import '../../provider/experience_points.dart';
-import '../../provider/task.dart';
-import '../../provider/tasks.dart';
-import '../../provider/charts_stats.dart';
+import '../../../provider/experience_points.dart';
+import '../../../provider/task.dart';
+import '../../../provider/tasks.dart';
+import '../../../provider/charts_stats.dart';
 
 // Services Imports
-import '../../services/notifications.dart';
+import '../../../services/notifications.dart';
 
 class ScheduledCard extends StatelessWidget {
   Widget getIcon(Difficulty difficulty) {
